+++ migrations/2021_02_07_add_max_post_content_length_column.php
@@ -14,13 +14,18 @@ use Illuminate\Database\Schema\Builder;
 
 return [
     'up' => function (Builder $schema) {
+        if ($schema->hasColumn('webhooks', 'max_post_content_length')) {
+            return;
+        }
         $schema->table('webhooks', function (Blueprint $table) {
             $table->unsignedInteger('max_post_content_length')->nullable();
         });
     },
     'down' => function (Builder $schema) {
-        $schema->table('webhooks', function (Blueprint $table) {
-            $table->dropColumn('max_post_content_length');
-        });
+        if ($schema->hasColumn('webhooks', 'max_post_content_length')) {
+            $schema->table('webhooks', function (Blueprint $table) {
+                $table->dropColumn('max_post_content_length');
+            });;
+        }
     },
 ];
